Rename the deploy-[development|production]-vpc scripts to deploy-[development|production]
According to Kevin:

"I _think_ the intention was to have standard NPM commands that you could copy and paste into a package.json:
if you need to be in a VPC, use the VPC deployment commands. If not, use the other commands."

They were boilerplate all-purpose scripts.
Since this app uses a DB connection and needs to be on a vpc..and the .travis.yml file refers to the -vpc scripts..
I've renamed the '-vpc' scripts, and references to them to drop the '-vpc'.
I've also removed the non-vpc variations since we'll ALWAYS want these on a VPC (db access).
This should make the package.json more self-evident.